Separate logging and auth from main code
